// types/tagTypes.ts
export enum TagTypes {
  user = "user",
  collection = "collection",
  subCategories = "sunCategories",
  brand = "brand",
  banner = "banner",
  color = "color",
  size = "size",
  product = "product",
  order = "order",
  role = "role",
  coupon = "coupon"
}

export const TagTypeList = [
  TagTypes.user,
  TagTypes.collection,
  TagTypes.subCategories,
  TagTypes.brand,
  TagTypes.banner,
  TagTypes.color,
  TagTypes.size,
  TagTypes.product,
  TagTypes.order,
  TagTypes.role,
  TagTypes.coupon
];